minor clean ups of result panel
