+++ .changeset/silver-ghosts-compare.md
@@ -1,0 +1,3 @@
+---
+'@finos/legend-query-builder': patch
+---
